+++ src/pytest_report_extras/plugin.py
@@ -67,9 +67,10 @@ def _fx_screenshots(request):
 
 
 @pytest.fixture(scope="session")
-def _fx_report_html(request):
+def _fx_html(request):
     """ The folder storing the pytest-html report """
-    return utils.get_folder(request.config.getoption("--html", default=None))
+    path = utils.get_folder(request.config.getoption("--html", default=None))
+    return path if utils.is_package_installed("pytest-html") else None
 
 
 @pytest.fixture(scope="session")
@@ -79,9 +80,10 @@ def _fx_single_page(request):
 
 
 @pytest.fixture(scope="session")
-def _fx_report_allure(request):
-    """ Whether the allure-pytest plugin is being used """
-    return request.config.getoption("--alluredir", default=None)
+def _fx_allure(request):
+    """ The folder storing the allure report """
+    path = request.config.getoption("--alluredir", default=None)
+    return path if utils.is_package_installed("allure-pytest") else None
 
 
 @pytest.fixture(scope="session")
@@ -96,18 +98,12 @@ def _fx_sources(request):
     return request.config.getini("extras_sources")
 
 
-@pytest.fixture(scope='session')
-def _fx_check(_fx_report_html, _fx_report_allure, _fx_single_page):
-    """ Verifies preconditions before using this plugin. """
-    utils.check_options(_fx_report_html, _fx_report_allure)
-
-
 #
 # Test fixture
 #
 @pytest.fixture(scope="function")
-def report(_fx_report_html, _fx_single_page, _fx_screenshots, _fx_sources, _fx_indent, _fx_report_allure, _fx_check):
-    return Extras(_fx_report_html, _fx_single_page, _fx_screenshots, _fx_sources, _fx_indent, _fx_report_allure)
+def report(_fx_html, _fx_single_page, _fx_screenshots, _fx_sources, _fx_indent, _fx_allure):
+    return Extras(_fx_html, _fx_single_page, _fx_screenshots, _fx_sources, _fx_indent, _fx_allure)
 
 
 #
@@ -116,13 +112,13 @@ def report(_fx_report_html, _fx_single_page, _fx_screenshots, _fx_sources, _fx_i
 
 # Global variables to store required fixtures to handle tms and issue markers
 # Workaround for https://github.com/pytest-dev/pytest/issues/13101
-fx_html = None
 fx_allure = None
-fx_tms_link_pattern = None
+fx_html = None
 fx_issue_link_pattern = None
 fx_links_column = "all"
 fx_single_page = False
 fx_title = ""
+fx_tms_link_pattern = None
 
 
 @pytest.hookimpl(hookwrapper=True)
@@ -130,7 +126,11 @@ def pytest_runtest_makereport(item, call):
     """
     Complete pytest-html report with extras and Allure report with attachments.
     """
-    global fx_html, fx_allure, fx_single_page, fx_issue_link_pattern, fx_tms_link_pattern, fx_links_column
+    global fx_allure
+    global fx_html
+    global fx_issue_link_pattern
+    global fx_links_column
+    global fx_tms_link_pattern
 
     outcome = yield
     pytest_html = item.config.pluginmanager.getplugin("html")
@@ -146,7 +146,7 @@ def pytest_runtest_makereport(item, call):
         call.when == "setup" and
         (report.failed or report.skipped) and
         fx_html is not None and pytest_html is not None and
-        "report" in item._fixtureinfo.argnames
+        "report" in item.fixturenames  # _fixtureinfo.argnames
     ):
         if report.failed:
             status = Status.ERROR
@@ -155,7 +155,18 @@ def pytest_runtest_makereport(item, call):
         header = decorators.get_header_rows(item, call, report, links, status)
         extras.append(pytest_html.extras.html(f'<table class="extras_header">{header}</table>'))
 
-    # Exit if the test is not using the 'report' fixtures
+    # Add 'report' fixture for pytest-bdd tests
+    if (
+        report.when == "call" and
+        fx_html and pytest_html is not None and
+        item.config.pluginmanager.has_plugin("pytest-bdd") and
+        "request" in item.fixturenames and
+        "_pytest_bdd_example" in item.fixturenames
+    ):
+        item.funcargs["report"] = Extras(_fx_html, _fx_single_page, _fx_screenshots, _fx_sources, _fx_indent, _fx_allure)
+        item.fixturenames.append("report")
+
+    # Exit if the test is not using the 'report' fixture
     if not ("request" in item.funcargs and "report" in item.funcargs):
         report.extras = extras  # add links to the report before exiting
         return
@@ -166,7 +177,13 @@ def pytest_runtest_makereport(item, call):
         try:
             feature_request = item.funcargs["request"]
             fx_report = feature_request.getfixturevalue("report")
+            if (
+                "_pytest_bdd_example" in item.fixturenames and 
+                len(fx_report.comments) + len(fx_report.attachments) + len(fx_report.multimedia) == 0
+            ):
+                return
             fx_screenshots = feature_request.getfixturevalue("_fx_screenshots")
+            fx_single_page = feature_request.getfixturevalue("_fx_single_page")
             target = fx_report.target
         except pytest.FixtureLookupError as error:
             utils.log_error(report, "Could not retrieve test fixtures", error)
@@ -283,15 +300,26 @@ def pytest_configure(config):
     """
     Performs setup actions and sets global variables.
     """
-    global fx_html, fx_allure, fx_issue_link_pattern, fx_tms_link_pattern, fx_single_page, fx_title, fx_links_column
-    # Retrieve some options
-    fx_html = utils.get_folder(config.getoption("--html", default=None))
+    global fx_allure
+    global fx_html
+    global fx_issue_link_pattern
+    global fx_links_column
+    global fx_single_page
+    global fx_title
+    global fx_tms_link_pattern
+
+    # Initialize global variables
     fx_allure = config.getoption("--alluredir", default=None)
-    fx_single_page = config.getoption("--self-contained-html", default=False)
-    fx_tms_link_pattern = config.getini("extras_tms_link_pattern")
+    fx_html = utils.get_folder(config.getoption("--html", default=None))
     fx_issue_link_pattern = config.getini("extras_issue_link_pattern")
     fx_links_column = config.getini("extras_links_column")
+    fx_single_page = config.getoption("--self-contained-html", default=False)
     fx_title = config.getini("extras_title")
+    fx_tms_link_pattern = config.getini("extras_tms_link_pattern")
+    if not utils.is_package_installed("pytest-html"):
+        fx_html = None
+    if not utils.is_package_installed("allure-pytest"):
+        fx_allure = None
 
     # Add markers
     config.addinivalue_line("markers", "issue(keys, icon): The list of issue keys to add as links")
@@ -308,7 +336,7 @@ def pytest_configure(config):
 
 @pytest.hookimpl(tryfirst=True)
 def pytest_sessionstart(session):
-    """ Create report asserts. """
+    """ Create report assets. """
     global fx_html, fx_single_page
     if fx_html is not None:
         utils.create_assets(fx_html, fx_single_page)
@@ -320,9 +348,10 @@ def pytest_sessionfinish(session, exitstatus):
     global fx_html
     if fx_html is not None:
         utils.delete_empty_subfolders(fx_html)
+    utils.check_options(fx_html, fx_allure)
 
 
-if importlib.util.find_spec("html") is not None:
+if utils.is_package_installed("pytest-html"):
     def pytest_html_report_title(report):
         global fx_title
         report.title = fx_title